+++ src/Blink.h
@@ -7,7 +7,7 @@ class Blink : public Animation {
 public:
     Blink(const RGB& colorCurrent, const RGB& colorTarget, const Point& point);
 
-    bool update(int currentTime) override;
+    bool update(unsigned long currentTime) override;
 
 private:
     Waiter waiter;
